p2p-webrtc/voice-agent/requirements.txt
python-dotenv
fastapi[all]
uvicorn
aiohttp
pipecat-ai[google,silero,webrtc]>=0.0.82